Revert "feat: Money의 인스턴스 필드를 BigInteger로 변환"
This reverts commit e4ff251673bca596c65733c5d5691fb41252843e.

+++ src/main/java/lotto/exceptions/PurchaseMoneyIllegalArgumentException.java
@@ -1,14 +1,12 @@
 package lotto.exceptions;
 
-import java.math.BigInteger;
-
 public class PurchaseMoneyIllegalArgumentException extends IllegalArgumentException {
 	public static final String MESSAGE
 			= "는 유효한 금액이 아닙니다. 금액은 0보다 큰 1000의 배수여야 합니다.";
 
-	private final BigInteger invalidPurchaseMoney;
+	private final int invalidPurchaseMoney;
 
-	public PurchaseMoneyIllegalArgumentException(final BigInteger invalidPurchaseMoney) {
+	public PurchaseMoneyIllegalArgumentException(final int invalidPurchaseMoney) {
 		super(invalidPurchaseMoney + MESSAGE);
 		this.invalidPurchaseMoney = invalidPurchaseMoney;
 	}
